Add :trytiny_var option for $_ default exception
This proposed change allows (lexically) setting $_ as the default
exception object for kiss, hug, barf, and bleep. This makes life easier
when using Ouch in team with Try::Tiny, because the latter uses $_ as
the exception object that is passed around.

So, instead of doing either of the following:

    use Try::Tiny;
    use Ouch;
    try { ouch 500, 'OUCH!' }
    catch {
       if (kiss 500, $_) { ... }
       else { ... }
    };

    try { ouch 404, 'No luck' }
    catch {
       local $@ = $_;
       if (kiss 404) { ... }
       else { ... }
    }

it's possible to "just" do this:

    use Try::Tiny;
    use Ouch qw<:trytiny_var>;
    try { ouch 500, 'OUCH!' }
    catch {
       if (kiss 500) { ... }
       else { ... }
    };

i.e. get full Ouch laziness and DWIM-mery back.

The change also includes tests for the new feature and documentation.